Adds base layout for the application
Implements the base layout with a blue bar and content container for the application.

It introduces a wrapper with a blue bar for visual appeal and a content container for the main content with a defined maximum width, padding, and background color.

The background color is moved to the html and body to cover the whole screen.