Fixes authentication and authorization issues
Updates the authentication scheme to use the 'oid' claim for user identification.

This change addresses issues where user authentication and authorization were failing due to incorrect claim mapping. The 'preferred_username' claim was replaced with 'oid' to align with the actual claim used for user identification. This ensures that user emails are correctly retrieved and used for authorization checks.

Also updates Azure AD ClientId and Domain in docker compose.