Refactors ProjectExpandable component to use slots
Moves the project stats and task list into slots within the
ProjectExpandable component.

This change improves flexibility and allows for easier customization
of the component's content. Removes the week property, DayPillStrip
and TaskStrip from the component and passes them through the slot in
the TimeTracker component.

Also uses a chevron-up icon when the project is open and chevron-down
when it's closed.