1010 valgfri rekkeflge p prosjekter i ti (#1036)
* Improves task filter input appearance

Enhances the visual appearance of the task filter input by adjusting padding.

* Enables reordering of favorite projects

Adds the ability for users to reorder their favorite projects
in the time tracker view via drag and drop, enhancing
customization and workflow efficiency.

Includes the SortableJS library and supporting components
to handle the drag-and-drop functionality.

* Enables project reordering in Time Tracker

Allows users to reorder favorite projects in the Time Tracker.

This introduces a drag-and-drop interface within the project sorter, updating the order of projects in local storage and reflecting the change in the UI. It enhances user experience by providing a way to prioritize projects based on individual workflow.