Feature/show value of timebank (#1053)
* Refactors local storage handling and fixes icon import

Renames and expands the local project composable to a more generic local storage composable.

This change also introduces salary persistence, allowing users to save and retrieve salary information.
Updates icon import path in the error box component.

* Adds time bank settings and value display

Implements a modal for time bank settings, allowing users to set their salary and calculate the monetary value of their time bank.

The value is calculated based on the user's provided salary and displayed in the time bank overview.

* Adds save button to settings modal

Adds a "Save and close" button to the settings modal.
This allows users to confirm and apply their
settings changes before closing the modal.

Also adds styling for the new button

* Enables closing modal by clicking outside

Allows users to close the modal by clicking on the overlay area.

This provides a more intuitive user experience by allowing users
to dismiss the modal without having to click the close button.