Enhances status bar with weekly hours summary (#1054)
* Enhances status bar with weekly hours summary

Replaces the static weekly hours display in the status bar with a new component that shows total hours for the week and a hoverable expander.

The expander displays hours logged for each day of the current week. The component calculates and sorts the hours per day based on time entries.

* Removes obsolete TODO comment

Removes a TODO comment that is no longer relevant.

* Redo expander logic to support legacy browsers such as Firefox and Safari

* Uses Norwegian locale for date formatting

Updates the date formatting in the invoice rate visualizer to use the Norwegian locale ("nb-NO") to ensure correct date representation for Norwegian users.