Persists project sort order to database (#1108)
* Persists project sort order to database

Saves the order of favorite projects to the database when the user reorders them in the project sorter.

This ensures that the project order is consistent across sessions and devices.
The index is added to the projects and then passed to the api to persist in the database.

* Removes unnecessary console log

Removes a console log statement from the `setLocalProjects` function.
The console log was used for debugging purposes and is no longer needed.

* Refactor

* Handles errors when updating project order

Adds a try-catch block to handle potential errors during the
project favorite order update process. This prevents the
application from crashing and logs the error for debugging.